Fix database path issues
